Moved page objects into proposed new structure.
