+++ pages/desktop/regions/__init__.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+
 # ***** BEGIN LICENSE BLOCK *****
 # Version: MPL 1.1/GPL 2.0/LGPL 2.1
 #
@@ -12,14 +13,15 @@
 # for the specific language governing rights and limitations under the
 # License.
 #
-# The Original Code is Firefox Input.
+# The Original Code is Mozilla WebQA Tests.
 #
-# The Initial Developer of the Original Code is
-# Mozilla Corp.
+# The Initial Developer of the Original Code is Mozilla.
+#
 # Portions created by the Initial Developer are Copyright (C) 2011
 # the Initial Developer. All Rights Reserved.
 #
-# Contributor(s): Dave Hunt <dhunt@mozilla.com>
+# Contributor(s):
+#   Dave Hunt <dhunt@mozilla.com>
 #
 # Alternatively, the contents of this file may be used under the terms of
 # either the GNU General Public License Version 2 or later (the "GPL"), or
@@ -34,12 +36,3 @@
 # the terms of any one of the MPL, the GPL or the LGPL.
 #
 # ***** END LICENSE BLOCK *****
-'''
-Created on Jan 28, 2011
-'''
-import input_base_page
-
-
-class ThanksPage(input_base_page.InputBasePage):
-
-    _page_title = u'Thanks :: Firefox Input'
