+++ pages/desktop/submit_feedback.py
@@ -40,11 +40,10 @@
 '''
 Created on Jan 28, 2011
 '''
-import input_base_page
-import thanks_page
+from pages.base import InputBasePage
 
 
-class SubmitFeedbackPage(input_base_page.InputBasePage):
+class SubmitFeedbackPage(InputBasePage):
 
     _page_title = 'Submit Feedback :: Firefox Input'
 
@@ -127,6 +126,7 @@ class SubmitFeedbackPage(input_base_page.InputBasePage):
         self.selenium.click(self._submit_feedback_locator)
         self.selenium.wait_for_page_to_load(self.timeout)
         if expected_result == 'success':
-            return thanks_page.ThanksPage(self.testsetup)
+            from pages.desktop.thanks import ThanksPage
+            return ThanksPage(self.testsetup)
         else:
             self.wait_for_element_visible(self._error_locator)
