+++ pages/desktop/submit_idea.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
 # ***** BEGIN LICENSE BLOCK *****
 # Version: MPL 1.1/GPL 2.0/LGPL 2.1
 #
@@ -19,9 +20,8 @@
 # Portions created by the Initial Developer are Copyright (C) 2011
 # the Initial Developer. All Rights Reserved.
 #
-# Contributor(s): Vishal
-#                 David Burns
-#                 Dave Hunt <dhunt@mozilla.com>
+# Contributor(s): Dave Hunt <dhunt@mozilla.com>
+#                 Bob Silverberg <bob.silverberg@gmail.com>
 #                 Bebe <florin.strugariu@softvision.ro>
 #
 # Alternatively, the contents of this file may be used under the terms of
@@ -38,57 +38,29 @@
 #
 # ***** END LICENSE BLOCK *****
 '''
-Created on Nov 24, 2010
+
+Created on Dec 22, 2010
+
 '''
-from page import Page
-import input_base_page
-import product_filter_region
+from pages.desktop.submit_feedback import SubmitFeedbackPage
 
 
-class SitesPage(input_base_page.InputBasePage):
+class SubmitIdeaPage(SubmitFeedbackPage):
 
-    _page_title = 'Sites :: Firefox Input'
+    _feedback_locator = 'id=idea-description'
+    _remaining_character_count_locator = 'css=#idea-description-counter'
+    _submit_feedback_locator = 'css=#idea .submit span'
+    _error_locator = 'css=#idea .errorlist li'
+    _back_locator = 'css=#idea > header > nav > a'
 
-    _sites_locator = "id('themes')//li[@class='site']"
-
-    def go_to_sites_page(self):
-        self.selenium.open('/sites/')
+    def go_to_submit_idea_page(self):
+        self.selenium.open('/feedback#idea')
         self.is_the_current_page
+        self.wait_for_element_visible(self._idea_page_locator)
 
     @property
-    def product_filter(self):
-        return product_filter_region.ProductFilter.ComboFilter(self.testsetup)
+    def is_submit_feedback_enabled(self):
+        return not self.selenium.is_element_present('css=#idea .submit a.disabled')
 
-    @property
-    def site_count(self):
-        return int(self.selenium.get_xpath_count(self._sites_locator))
-
-    @property
-    def sites(self):
-        return [self.Site(self.selenium, i + 1) for i in range(self.site_count)]
-
-    def site(self, index):
-        return self.Site(self.testsetup, index)
-
-    class Site(Page):
-
-        _name_locator = " .name a"
-
-        def __init__(self, testsetup, index):
-            Page.__init__(self, testsetup)
-            self.index = index
-
-        def absolute_locator(self, relative_locator):
-            return self.root_locator + relative_locator
-
-        @property
-        def root_locator(self):
-            return "css=#themes .site:nth(%s)" % (self.index - 1)
-
-        @property
-        def name(self):
-            return self.selenium.get_text(self.absolute_locator(self._name_locator))
-
-        def click_name(self):
-            self.selenium.click(self.absolute_locator(self._name_locator))
-            self.selenium.wait_for_page_to_load(self.timeout)
+    def is_visible(self):
+        return self.selenium.is_visible(self._idea_page_locator)
